Created container for python shelter scraper server.
