+++ api_container/requirements.txt
@@ -1,0 +1,5 @@
+dotenv
+supabase
+fastapi
+requests
+uvicorn
