--- scripts/requirements.txt
@@ -1,55 +1,0 @@
-annotated-doc==0.0.4
-annotated-types==0.7.0
-anyio==4.13.0
-cachetools==6.2.6
-certifi==2026.4.22
-cffi==2.0.0
-charset-normalizer==3.4.7
-click==8.3.3
-cryptography==48.0.0
-deprecation==2.1.0
-fastapi==0.136.1
-fsspec==2026.4.0
-h11==0.16.0
-h2==4.3.0
-hpack==4.1.0
-httpcore==1.0.9
-httpx==0.28.1
-hyperframe==6.1.0
-idna==3.13
-markdown-it-py==4.2.0
-mdurl==0.1.2
-mmh3==5.2.1
-multidict==6.7.1
-packaging==26.2
-postgrest==2.30.0
-propcache==0.4.1
-pycparser==3.0
-pydantic==2.13.4
-pydantic_core==2.46.4
-Pygments==2.20.0
-pyiceberg==0.11.1
-PyJWT==2.12.1
-pyparsing==3.3.2
-pyroaring==1.1.0
-python-dateutil==2.9.0.post0
-python-dotenv==1.2.2
-realtime==2.30.0
-requests==2.33.1
-rich==14.3.4
-six==1.17.0
-soupsieve==2.8.3
-starlette==1.0.0
-storage3==2.30.0
-StrEnum==0.4.15
-strictyaml==1.7.3
-supabase==2.30.0
-supabase-auth==2.30.0
-supabase-functions==2.30.0
-tenacity==9.1.4
-typing-inspection==0.4.2
-typing_extensions==4.15.0
-urllib3==2.7.0
-websockets==15.0.1
-yarl==1.23.0
-zstandard==0.25.0
